Seperate server routes into services and controllers. Clean up server.
