+++ README.md
@@ -1,6 +1,6 @@
 
 # Pandash
-
+CURRENTLY IN DEVELOPMENT
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 1.0.0.
 
 ## Why?
