+++ src/server/services/cities.service.js
@@ -1,0 +1,18 @@
+let request = require('request'),
+    Q       = require('q');
+
+let citiesService = {};
+
+citiesService.getData = getData;
+
+module.exports = citiesService;
+
+function getData (endpoint) {
+    let def = Q.defer();
+
+    request(endpoint, function(err, response, body){
+        def.resolve(JSON.parse(body));
+    });
+
+    return def.promise;
+}
